requirements.txt
asgiref==3.5.1
certifi==2021.10.8
chardet==3.0.4
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pytz==2022.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
tzdata==2022.1
